// src/payload-types.ts
export interface Media {
  id: number;
  alt: string;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
// ...
export interface Project {
  id: number;
  title: string;
  /**
   * URL-safe identifier
   */
  slug: string;
  /**
   * One-liner for list view
   */
  summary: string;
  description?: {
    root: {
      type: string;
      children: {
        type: any;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  } | null;
  role?: string | null;
  outcomes?:
    | {
        outcome: string;
        id?: string | null;
      }[]
    | null;
  techStack?:
    | {
        tech: string;
        id?: string | null;
      }[]
    | null;
  heroImage?: (number | null) | Media;
  gallery?:
    | {
        image: number | Media;
        id?: string | null;
      }[]
    | null;
  externalUrl?: string | null;
  featured?: boolean | null;
  date?: string | null;
  sortOrder?: number | null;
  updatedAt: string;
  createdAt: string;
}